# migration-orchestration/contracts/kalel-openapi.yaml
openapi: 3.1.0
info:
  title: Kal El Content API — Máquina Nerd boundary
  version: 1.0.0
  description: Contrato alvo. O adapter Next pode mapear a API Kal El existente para este modelo.
servers:
  - url: https://kalel.example.invalid/api/v1
security:
  - serviceBearer: []
paths:
...
  /content/sitemap:
    get:
      operationId: sitemap
      parameters:
        - { name: kind, in: query, required: true, schema: { enum: [articles, categories, authors, tags] } }
        - { name: cursor, in: query, schema: { type: string } }
      responses: { '200': { description: Sitemap page } }
...
components:
  securitySchemes: { serviceBearer: { type: http, scheme: bearer } }
...